Fixed some bugs in Tree generation with TOML load
